Add db-modules explicitly to projects for reference
Swap from wildcard -> building the file list automatically to GeneratedFileList.props to be imported by
db-modules.vcxproj to load.

This is because VS caches the file list; it would behave fine for building, but effectively in a clean environment
what would happen BEFORE is this:

1. Solution loaded for the first time. No dependencies fetched/built yet.
2. Build initiated. Submodules are fetched, including db-modules. db-modules project won't update visually.
3. Builds completes successfully, db-modules project isn't updated visually (as expected).
4. Solution restarted (or db-modules reloaded) -- db-modules project STILL not updated, despite the files existing.
5. .vs folder deleted and solution restarted again -- db-modules project updated.

It caching this is very annoying, so instead, on build we generate the file list explicitly.
This isn't cached.

So behaviour now, in a clean environment is:

1. Solution loaded for the first time. No dependencies fetched/built yet.
2. Build initiated. Submodules are fetched, including db-modules. db-modules project won't update visually.
3. Builds completes successfully, db-modules project isn't updated visually (as expected).
4. Solution restarted (or db-modules reloaded) -- db-modules project updated visually.

This is expected behaviour.

Ideally we'd be able to invoke a visual reload of the project but I think for this to work,
we'd have to change things around a bit with our dependency management.

It's not critical for use and most wouldn't even notice it anyway when first looking at the project.
Most likely a restart would've happened naturally before even caring what's in this project.