Add db-library wrapper code
This effectively, currently just wraps our database connection management
and recordset logic.

It implements the logic necessary to take a given generated model and fetch
all associated columns, binding them to the model as in the generated model code.

Additionally it can enforce further filtering logic via SqlBuilder if need be.

As far as database configuration goes, due to the way it currently works (which should
be cleaned up), the model associates with a specific database via a string identified (e.g. GAME).
The server must manage a DatabaseConnManager instance and supply it with required configuration
for the various types it actually uses, e.g. (this would be from config file, though):
    mgr.SetDatasourceConfig("GAME", "OpenKO", "knight", "knight");

For now this is primitive - it still needs proper pooling/persistent connections.
As far as the recordset code is concerned however, it can establish a new connection
for each recordset as MFC was already doing. This is fine.

It'll need to be changed for the sake of the stored procedures though.