Update .clang-format with documentation
Add a specific override for .clang-format for the client code (WarFare).
This is... temporary, until code can be refactored.
There's no reason for lines to be so long normally, but it has some ridiculously long-lined code
which is better left together (otherwise it breaks it up horrendously).
The code should be refactored and this override removed.

Add .clang-format "bridges" to chain between directories.
These *shouldn't* be necessary (because documentation states it checks the parent, and then its parent), but for the client's override,
it still seems to use fallback code instead somehow. Directly chaining this seems to cause proper inheritance.