Add DB_COMPAT_PADDED_NAMES
This allows for databases where the names are padded (due to being stored in CHAR(N) columns instead of VARCHAR(N)).
Our existing database has a few such cases, combined with some data remnants from when a column was padded but is no longer.

This will be entirely replaced in future, but for now we'll just allow for such cases.

Closes #479