Server: More informative listen/connect behaviour
The servers will now report the main port they're listening on and when it
allows connections (it can listen, but block until connections are allowed).

Additionally, regarding AI, we report and log the IP:port we're connecting to.

Removed the awkward ResetOutputList() behaviour so we can log appropriately.
This will all be changed with a console at some point, so this really doesn't
make sense to keep anyway.

Resolves #482