Client: Add connection error debug for debug builds
It has been quite common for users to be confused about where the client is connecting to
and how it's configured.

This change overrides connection error messages to provide more obvious info on where it's connecting,
what error code we got, and where exactly this IP was configured (the ports are hardcoded).